+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/Modules/AzMarkers.cs
@@ -61,14 +61,12 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask Add(Map map, IEnumerable<HtmlMarker> items)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
         }
 
         public async ValueTask Clear(Map map)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference);
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference);
         }
 
         public async ValueTask Remove(Map map, HtmlMarker item)
@@ -78,8 +76,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask Remove(Map map, IEnumerable<HtmlMarker> items)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
         }
 
         private static string GetJsInteropMethod([CallerMemberName] string name = "")