+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/Modules/AzPopups.cs
@@ -115,15 +115,13 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask<List<IMapObjectReference>> Add(Map map, IEnumerable<Popup> items, bool getReferences = false)
         {
-            var module = await moduleTask.Value;
-            var results = await module.InvokeAsync<List<MapObjectReference>>(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
+            var results = await moduleTask.InvokeAsyncInternal<List<MapObjectReference>>(GetJsInteropMethod(), map.MapReference, items?.Cast<object>().ToList());
             return [.. results.Cast<IMapObjectReference>()];
         }
 
         public async ValueTask<IMapObjectReference?> AddHoverPopup(Map map, ILayer layer, Popup item, IEnumerable<string> placeholders, bool getReference = false)
         {
-            var module = await moduleTask.Value;
-            return await module.InvokeAsync<MapObjectReference?>(GetJsInteropMethod(), map.MapReference, layer.Id, item, placeholders, getReference);
+            return await moduleTask.InvokeAsyncInternal<MapObjectReference?>(GetJsInteropMethod(), map.MapReference, layer.Id, item, placeholders, getReference);
         }
 
         public async ValueTask<IMapObjectReference> GetReference(Map map, Popup item)
@@ -134,8 +132,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask<List<IMapObjectReference>> GetReferences(Map map, IEnumerable<Popup> items)
         {
-            var module = await moduleTask.Value;
-            var results = await module.InvokeAsync<List<MapObjectReference>>(GetJsInteropMethod(), map.MapReference, items);
+            var results = await moduleTask.InvokeAsyncInternal<List<MapObjectReference>>(GetJsInteropMethod(), map.MapReference, items);
             return [.. results.Cast<IMapObjectReference>()];
         }
 
@@ -148,8 +145,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask Remove(Map map, IEnumerable<Popup> items)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, items);
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, items);
         }
 
         public async ValueTask Remove(Map map, IMapObjectReference item)
@@ -159,8 +155,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
 
         public async ValueTask Remove(Map map, IEnumerable<IMapObjectReference> items)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, items.Select(e => e.JsReference).ToList());
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, items.Select(e => e.JsReference).ToList());
             await items.DisposeItems();
         }