+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/Modules/Events/AzStyleControlEvents.cs
@@ -43,20 +43,17 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules.Events
 
         public async ValueTask Add<T>(DotNetObjectReference<T> dotNetRef, Map map, IEnumerable<StyleControlEvent> events) where T : class
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), dotNetRef, map.MapReference, events.Cast<object>().ToList());
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), dotNetRef, map.MapReference, events.Cast<object>().ToList());
         }
 
         public async ValueTask Remove(Map map, IEnumerable<StyleControlEvent> events)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, events.Cast<object>().ToList());
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, events.Cast<object>().ToList());
         }
 
         public async ValueTask RemoveBySource(Map map, IEnumerable<object> sources)
         {
-            var module = await moduleTask.Value;
-            await module.InvokeVoidAsync(GetJsInteropMethod(), map.MapReference, sources);
+            await moduleTask.InvokeVoidAsyncInternal(GetJsInteropMethod(), map.MapReference, sources);
         }
 
         private static string GetJsInteropMethod([CallerMemberName] string name = "")