+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/scripts/atlas/Sources.ts
@@ -1,68 +1,64 @@
 import * as atlas from "azure-maps-control"
 import { MapObjectReference } from "./common"
 import { Helpers } from "./common/Helpers";
+import { Logger, LogLevel } from "./common/Logger";
 
 export class Sources {
-    public static add(map: atlas.Map, mapSources: MapSource[], getReferences: boolean = false) {
+    public static add(map: atlas.Map, sources?: MapSource[], getReferences: boolean = false): MapObjectReference[] {
+        const eventName = "Sources.add";
         const results: MapObjectReference[] = [];
         const mapId = Helpers.getMapId(map);
 
-        mapSources ?? [];
+        if (Helpers.hasNoElements(sources))
+            return results;
 
-        mapSources.forEach((ms) => {
-            let ds: atlas.source.Source | undefined = undefined;
+        if (Helpers.hasDuplicates(sources, "id", true, true))
+            return results; //Will never reach here if true.
 
-            switch (ms.type) {
+        const existing = map.sources.getSources(); 
+
+        sources!.forEach((source) => {
+            let ds = this.#getInstance(map, source, existing);
+            if (ds) {
+                Logger.logMapMessage(mapId, LogLevel.Warn, `${eventName}: Source already exists.`, source);
+                return; //continue
+            }
+
+            switch (source.type) {
                 case SourceType.Data:
-                    ds = new atlas.source.DataSource(ms.id, ms.options);
+                    ds = new atlas.source.DataSource(source.id, source.options);
                     break;
                 case SourceType.ElevationTile:
-                    ds = new atlas.source.ElevationTileSource(ms.id, ms.options);
+                    ds = new atlas.source.ElevationTileSource(source.id, source.options);
                     break;
                 case SourceType.VectorTile:
-                    ds = new atlas.source.VectorTileSource(ms.id, ms.options);
+                    ds = new atlas.source.VectorTileSource(source.id, source.options);
                     break;
                 default:
             }
 
             if (ds) {
                 map.sources.add(ds);
-
                 if (getReferences)
-                    results.push(this.#getReference(mapId, ds, ms.id));
+                    results.push(this.#getReference(mapId, ds, source.id));
             }
         });
 
         return results;
     }
 
-    public static getSources(map: atlas.Map, ids?: string[]) {
-        if (ids) {
-            return this.#getSourcesById(map, ids);
+    public static getReferences(map: atlas.Map, sources?: any[]): MapObjectReference[] {
+        const results: MapObjectReference[] = [];
+        const mapId = Helpers.getMapId(map);
+        const existing = map.sources.getSources();
+
+        if (Helpers.hasNoElements(sources)) {
+            return results;
         }
-        else {
-            return this.#getSourcesAll(map);
-        } 
-    }
 
-    static #getSourcesAll(map: atlas.Map) {
-        const results: MapObjectReference[] = [];
-        const mapId = Helpers.getMapId(map);
-        const sources = map.sources.getSources();
-
-        sources.forEach((ds) => {
+        sources!.forEach((source) => {
+            const ds = this.#getInstance(map, source, existing);
             results.push(this.#getReference(mapId, ds));
-        });
-
-        return results;
-    }
-
-    static #getSourcesById(map: atlas.Map, ids: string[]) {
-        const results: MapObjectReference[] = [];
-        const mapId = Helpers.getMapId(map);
-
-        ids.forEach((srcId) => {
-            results.push(this.#getReference(mapId, map.sources.getById(srcId), srcId));
         });
 
         return results;
@@ -72,21 +68,14 @@ export class Sources {
         map.sources.remove(sources);
     }
 
-    static #getReference(mapId: string, ds: atlas.source.Source, id: string = "") {
+    static #getReference(mapId: string, ds?: atlas.source.Source, id?: string) {
         const type = this.#getSourceType(ds);
-        const dnr = !ds ? null : DotNet.createJSObjectReference(ds);
-        let dsId = ds?.getId();
-        if (Helpers.isEmptyOrNull(dsId)) {
-            dsId = id;
-        }
-        const msf: MapObjectReference = { mapId: mapId, id: dsId, type: type, jsReference: dnr };
-
-        return msf;
+        return Helpers.createMapObjectReference(mapId, type, ds, id ?? ds?.getId());
     }
 
-    static #getSourceType(ds?: atlas.source.Source | undefined): string | undefined {
+    static #getSourceType(ds?: atlas.source.Source | undefined): string {
         if (!ds)
-            return;
+            return "Undefined";
 
         if (ds instanceof atlas.source.DataSource)
             return SourceType.Data; 
@@ -95,7 +84,28 @@ export class Sources {
             return SourceType.ElevationTile; 
 
         if (ds instanceof atlas.source.VectorTileSource)
-            return SourceType.VectorTile; 
+            return SourceType.VectorTile;
+
+        return "Source";
+    }
+
+    static #getInstance(map: atlas.Map, source: any, sources: atlas.source.Source[], eventName?: string): atlas.source.Source | undefined {
+        let item: any;
+
+        if (source instanceof atlas.source.Source) {
+            item = source;
+        }
+        else {
+            const id = Helpers.getSourceId(source);
+            item = sources.findLast(item => Helpers.hasId(item, id));
+        }
+
+        if (!item && Helpers.isNotEmptyOrNull(eventName)) {
+            const mapId = Helpers.getMapId(map);
+            Logger.logMessage(mapId, LogLevel.Warn, `${eventName}: instance not found.`, source);
+        }
+
+        return item;
     }
 }