+++ src/MainLib/Sandbox/Components/Pages/Atlas/MapExtensions.cs
@@ -1,9 +1,9 @@
-﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Common;
-using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Configuration;
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Configuration;
 using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Controls;
 using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Events;
 using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Events.Payloads;
 using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Events.Types;
+using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Models.Exceptions;
 using Marqdouj.DotNet.Web.JsInterop.GeoJson;
 using Microsoft.FluentUI.AspNetCore.Components;
 using System.Text.Json;
@@ -24,6 +24,23 @@ namespace Sandbox.Components.Pages.Atlas
             internal string ToJsonMin()
             {
                 return JsonSerializer.Serialize(obj, jsonMinOptions);
+            }
+        }
+
+        /// <summary>
+        /// Checks if exception is <see cref="JS2CSharpException"/> and logs accordingly.
+        /// </summary>
+        /// <param name="logger"></param>
+        /// <param name="ex"></param>
+        public static void LogMapError(this ILogger logger, Exception ex)
+        {
+            if (ex is JS2CSharpException jsEx)
+            {
+                jsEx.LogError(logger);
+            }
+            else
+            {
+                logger.LogError(ex, null);
             }
         }